strange error fixed.
I switched to ruby 1.9.3 for heroku, and then (although my app started)
I got an error after signing up. Running the app on my local machine in
1.9.3 showed that the following code does not work in the 1.9.3 setup:

  'abcd'.chars.product('defg'.chars)

The error reports that the enumerable object doesn't have a product
method. However, I cannot see a difference between the 1.9.3 and 2.1.3
docs on that point.

At any rate, I rewrote that part of the code and it will now run
locally in 1.9.3. Let's see about heroku.